+++ packages/orq-rc/RELEASES.md
@@ -2029,3 +2029,13 @@ Based on:
 - [typescript v3.8.0-rc.7] packages/orq-rc
 ### Releases
 - [NPM v3.8.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.7 - packages/orq-rc
+
+## 2025-06-10 02:02:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.558.1 (2.623.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.8.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v3.8.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.8 - packages/orq-rc